requirements.txt
allure-pytest~=2.8.18
pytest~=6.1.1
selenium~=3.141.0
